pyproject.toml
[project]
name = "ai-audio-assistant-backend"
version = "0.1.0"
description = "AI Audio/Video Content Assistant Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "celery>=5.3.4",
    "greenlet>=3.3.0",
    "redis>=5.0.1",
    "minio>=7.2.0",
    "loguru>=0.7.2",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "tencentcloud-sdk-python>=3.0.1000",
    "yt-dlp>=2024.12.0",
    "aiohttp>=3.9.0",
    "python-multipart>=0.0.6",
    "passlib[bcrypt]>=1.7.4",
    "boto3>=1.34.0",
    "cos-python-sdk-v5>=1.9.32",
    "pyyaml>=6.0.0",
    "ujson>=5.9.0",
    "psycopg2-binary>=2.9.9",
    "email-validator>=2.1.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "mypy>=1.8.0",
    "black>=24.1.1",
    "isort>=5.13.2",
    "flake8>=7.0.0",
    "types-redis>=4.6.0",
    "types-pyyaml>=6.0.12",
]

# 删除 [build-system] 部分，不作为可安装包

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
